untested changes for user coins

+++ src/main/java/io/inaam/main/entity/UserCoin.java
@@ -3,7 +3,6 @@ package io.inaam.main.entity;
 import lombok.Data;
 
 import javax.persistence.*;
-import java.math.BigInteger;
 
 @Data
 @Entity
@@ -19,11 +18,7 @@ public class UserCoin
     @Column(name = "coin_id", nullable = false)
     private String coinId;
 
-//    @OneToOne
-//    @JoinColumn(name = "coin_id", referencedColumnName = "id", insertable = false, updatable = false)
-//    private Coin coin;
-
     @Basic
     @Column(name = "balance")
-    private BigInteger balance;
+    private int balance;
 }
